requirements.txt
# --- Core AI & Image Processing ---
ultralytics==8.1.30
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tensorflow-macos==2.15.0
numpy>=1.23.5,<1.25
opencv-python==4.8.1.78
Pillow==10.0.0

# --- Google Gemini AI ---
google-generativeai==0.1.0

# --- HTTP requests & date parsing ---
requests==2.31.0
python-dateutil==2.8.2

# --- API Server ---
fastapi==0.103.1
uvicorn==0.23.2
